+++ CHANGELOG.md
@@ -11,6 +11,8 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   of material property types (e.g. arrays of color buffers).
 - The model warper (V3) now supports chaining/mixing TPS warps/substitutions of meshes
   (previously, it was one or the other).
+- The model warper (V3) now supports `RecalculateWrapCylinderRadiusFromStationScalingStep`
+  for rescaling a `WarpCylinder`'s radius (#1013).
 
 
 ## [0.5.21] - 2025/02/28